+++ devices/TPS65386x-Q1/src/pmic_irq.c
@@ -569,7 +569,7 @@ static inline void IRQ_copyIrqStat(const Pmic_IrqStatus_t *src, Pmic_IrqStatus_t
 
 static inline void IRQ_setIntrStat(Pmic_IrqStatus_t *irqStat, uint32_t irqNum)
 {
-    const uint32_t mask = ((uint32_t)1U << (irqNum % PMIC_NUM_BITS_IN_INTR_ELEM));
+    const uint32_t mask = ((uint32_t)1UL << (irqNum % PMIC_NUM_BITS_IN_INTR_ELEM));
 
     if (irqNum < (uint32_t)PMIC_IRQ_NUM)
     {
@@ -582,7 +582,7 @@ static inline void IRQ_setIntrStat(Pmic_IrqStatus_t *irqStat, uint32_t irqNum)
 
 static inline void IRQ_clrIntrStat(Pmic_IrqStatus_t *irqStat, uint32_t irqNum)
 {
-    const uint32_t mask = ((uint32_t)1U << (irqNum % PMIC_NUM_BITS_IN_INTR_ELEM));
+    const uint32_t mask = ((uint32_t)1UL << (irqNum % PMIC_NUM_BITS_IN_INTR_ELEM));
 
     if (irqNum < (uint32_t)PMIC_IRQ_NUM)
     {
@@ -604,7 +604,7 @@ static uint8_t IRQ_getNextFlag(Pmic_IrqStatus_t *irqStat) {
         // For each bit in the element...
         for (bitPos = 0U; bitPos < PMIC_NUM_BITS_IN_INTR_ELEM; bitPos++) {
             // If the bit is set...
-            const uint32_t mask = (1U << bitPos);
+            const uint32_t mask = (uint32_t)(1UL << bitPos);
             if ((irqStat->intrStat[index] & mask) != 0U) {
                 // Clear bit in intrStat element and exit loop
                 irqStat->intrStat[index] &= ~mask;
@@ -802,7 +802,7 @@ static int32_t IRQ_handleRecordsForRegMask(const Pmic_Handle_t *handle,
         for (uint8_t i = 0U; (i < PMIC_IRQ_NUM) && (i < numCfgs); i++) {
             const uint8_t irqNum = cfgs[i].irqNum;
             const Pmic_IrqInfo_t *pIrq = &pmicIRQs[irqNum];
-            const uint8_t userMask = (uint8_t)(1U << pIrq->maskShift);
+            const uint8_t userMask = (uint8_t)(1UL << pIrq->maskShift);
 
             // If the current mask setting isn't targeted at the register we are
             // currently building, skip it
@@ -1230,7 +1230,7 @@ int32_t Pmic_irqClrAllFlags(const Pmic_Handle_t *handle) {
 
         // Clear OFF_STATE_STAT{1,2}_REG
         if (status == PMIC_ST_SUCCESS) {
-            status = Pmic_ioTxByte(handle, OFF_STATE_CLR_REG, 1U << OFF_STATE_STAT_CLR_SHIFT);
+            status = Pmic_ioTxByte(handle, OFF_STATE_CLR_REG, 1UL << OFF_STATE_STAT_CLR_SHIFT);
         }
 
         Pmic_criticalSectionStop(handle, PMIC_COMMUNICATION);
